fix #208
related to https://github.com/realodix/newt/commit/9077e9ae5756bf516dd6b8d25a310178a190aed9

+++ database/migrations/2018_08_03_114156_create_urls_table.php
@@ -15,7 +15,7 @@ class CreateUrlsTable extends Migration
     {
         Schema::create('urls', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('user_id')->nullable($value = true);
+            $table->unsignedBigInteger('user_id')->nullable($value = true);
             $table->string('url_key')->collation('utf8mb4_bin')->unique();
             $table->boolean('is_custom');
             $table->longText('long_url');
